Move integration tests helper methods to utils folder
