+++ test/integration/utils/voteClaim.js
@@ -24,30 +24,30 @@ async function voteClaim({ claimId, verdict, cl, cd, cr, voter }) {
   assert.equal(voteStatusAfter.toString(), '-1', 'voting should be closed');
 }
 
-async function acceptClaim({ staker, assessmentStakingAmount, as }) {
+async function acceptClaim({ staker, assessmentStakingAmount, as, assessmentId }) {
   const { payoutCooldownInDays } = await as.config();
   await as.connect(staker).stake(assessmentStakingAmount);
 
-  await as.connect(staker).castVotes([0], [true], ['Assessment data hash'], 0);
+  await as.connect(staker).castVotes([assessmentId], [true], ['Assessment data hash'], 0);
 
-  const { poll } = await as.assessments(0);
+  const { poll } = await as.assessments(assessmentId);
   const futureTime = poll.end + daysToSeconds(payoutCooldownInDays);
 
   await setTime(futureTime);
 }
 
-async function rejectClaim({ approvingStaker, rejectingStaker, as }) {
+async function rejectClaim({ approvingStaker, rejectingStaker, as, assessmentId }) {
   const assessmentStakingAmountForApproval = parseEther('1000');
   const assessmentStakingAmountForRejection = parseEther('2000');
   const { payoutCooldownInDays } = await as.config();
   await as.connect(approvingStaker).stake(assessmentStakingAmountForApproval);
 
-  await as.connect(approvingStaker).castVotes([0], [true], ['Assessment data hash'], 0);
+  await as.connect(approvingStaker).castVotes([assessmentId], [true], ['Assessment data hash'], 0);
 
   await as.connect(rejectingStaker).stake(assessmentStakingAmountForRejection);
-  await as.connect(rejectingStaker).castVotes([0], [false], ['Assessment data hash'], 0);
+  await as.connect(rejectingStaker).castVotes([assessmentId], [false], ['Assessment data hash'], 0);
 
-  const { poll } = await as.assessments(0);
+  const { poll } = await as.assessments(assessmentId);
   const futureTime = poll.end + daysToSeconds(payoutCooldownInDays);
 
   await setTime(futureTime);
